feat: restructure bank composables for modularity and backward compatibility
