+++ app/src/composables/bank.ts
@@ -1,551 +1,10 @@
-import { computed, ref, watch } from 'vue'
-import { useReadContract, useWriteContract, useWaitForTransactionReceipt, useAccount } from '@wagmi/vue'
-import { useQueryClient } from '@tanstack/vue-query'
-import { decodeFunctionData, getContract, type Address, parseEther, isAddress } from 'viem'
-import { config } from '@/wagmi.config'
-import { log, parseError } from '@/utils'
-import { useToastStore } from '@/stores'
-import { useTeamStore } from '@/stores'
-import BankABI from '@/artifacts/abi/bank.json'
-
 /**
- * Valid Bank contract function names extracted from ABI
+ * Bank contract composables - Legacy file
+ * 
+ * This file now re-exports from the modular bank composables structure
+ * for backward compatibility. New code should import from './bank/index.ts'
+ * or the specific modules directly.
  */
-export const BANK_FUNCTION_NAMES = {
-  // Read functions
-  PAUSED: 'paused',
-  OWNER: 'owner',
-  TIPS_ADDRESS: 'tipsAddress',
-  IS_TOKEN_SUPPORTED: 'isTokenSupported',
-  SUPPORTED_TOKENS: 'supportedTokens',
 
-  // Write functions
-  PAUSE: 'pause',
-  UNPAUSE: 'unpause',
-  CHANGE_TIPS_ADDRESS: 'changeTipsAddress',
-  CHANGE_TOKEN_ADDRESS: 'changeTokenAddress',
-  TRANSFER_OWNERSHIP: 'transferOwnership',
-  RENOUNCE_OWNERSHIP: 'renounceOwnership',
-  DEPOSIT_TOKEN: 'depositToken',
-  TRANSFER: 'transfer',
-  TRANSFER_TOKEN: 'transferToken',
-  SEND_TIP: 'sendTip',
-  SEND_TOKEN_TIP: 'sendTokenTip',
-  PUSH_TIP: 'pushTip',
-  PUSH_TOKEN_TIP: 'pushTokenTip',
-  INITIALIZE: 'initialize'
-} as const
-
-/**
- * Type for valid Bank contract function names
- */
-export type BankFunctionName = typeof BANK_FUNCTION_NAMES[keyof typeof BANK_FUNCTION_NAMES]
-
-/**
- * Validate if a function name exists in the Bank contract
- */
-export function isValidBankFunction(functionName: string): functionName is BankFunctionName {
-  return Object.values(BANK_FUNCTION_NAMES).includes(functionName as BankFunctionName)
-}
-
-/**
- * Legacy function - kept for backward compatibility
- */
-export function useBankGetFunction(bankAddress: string) {
-  const functionName = ref<string | undefined>()
-  const inputs = ref<string[] | undefined>([])
-  const args = ref<string[] | undefined>([])
-
-  async function getFunction(data: string): Promise<void> {
-    try {
-      const bank = getContract({
-        client: config.getClient(),
-        address: bankAddress as Address,
-        abi: BankABI
-      })
-      const func = decodeFunctionData({
-        abi: BankABI,
-        data: data as Address
-      })
-
-      functionName.value = func.functionName
-      args.value = func.args as string[]
-      inputs.value = bank.abi
-        .find((item) => item.name === func.functionName)
-        ?.inputs?.map((input) => input.name) as string[]
-    } catch (error) {
-      log.error(parseError(error))
-    }
-  }
-
-  return { execute: getFunction, data: functionName, inputs, args }
-}
-
-/**
- * Bank contract read operations
- */
-export function useBankReads() {
-  const teamStore = useTeamStore()
-  const bankAddress = computed(() => teamStore.getContractAddressByType('Bank'))
-  const isBankAddressValid = computed(() => !!bankAddress.value && isAddress(bankAddress.value))
-
-  const useBankPaused = () => {
-    return useReadContract({
-      address: bankAddress.value,
-      abi: BankABI,
-      functionName: BANK_FUNCTION_NAMES.PAUSED,
-      query: { enabled: isBankAddressValid } // This enable the query only if the bank address is available and valid
-    })
-  }
-
-  const useBankOwner = () => {
-    return useReadContract({
-      address: bankAddress.value,
-      abi: BankABI,
-      functionName: BANK_FUNCTION_NAMES.OWNER,
-      query: { enabled: isBankAddressValid }
-    })
-  }
-
-  const useBankTipsAddress = () => {
-    return useReadContract({
-      address: bankAddress.value,
-      abi: BankABI,
-      functionName: BANK_FUNCTION_NAMES.TIPS_ADDRESS,
-      query: { enabled: isBankAddressValid }
-    })
-  }
-
-  const useBankIsTokenSupported = (tokenAddress: Address) => {
-    return useReadContract({
-      address: bankAddress.value,
-      abi: BankABI,
-      functionName: BANK_FUNCTION_NAMES.IS_TOKEN_SUPPORTED,
-      args: [tokenAddress],
-      query: { enabled: computed(() => !!bankAddress.value && isAddress(bankAddress.value) && isAddress(tokenAddress)) }
-    })
-  }
-
-  const useBankSupportedTokens = (symbol: string) => {
-    return useReadContract({
-      address: bankAddress.value,
-      abi: BankABI,
-      functionName: BANK_FUNCTION_NAMES.SUPPORTED_TOKENS,
-      args: [symbol],
-      query: { enabled: computed(() => isBankAddressValid.value && !!symbol) }
-    })
-  }
-
-  return {
-    bankAddress,
-    isBankAddressValid,
-    useBankPaused,
-    useBankOwner,
-    useBankTipsAddress,
-    useBankIsTokenSupported,
-    useBankSupportedTokens
-  }
-}
-
-/**
- * Bank contract write operations with comprehensive error handling and query invalidation
- */
-export function useBankWrites() {
-  const { addSuccessToast, addErrorToast } = useToastStore()
-  const teamStore = useTeamStore()
-  const queryClient = useQueryClient()
-  const { chainId } = useAccount()
-  const bankAddress = computed(() => teamStore.getContractAddressByType('Bank'))
-  
-  // Store the current function name for query invalidation
-  const currentFunctionName = ref<BankFunctionName | null>(null)
-
-  const {
-    data: writeContractData,
-    writeContract,
-    isPending: isWritePending,
-    error: writeError
-  } = useWriteContract()
-
-  const {
-    data: receipt,
-    isLoading: isConfirming,
-    isSuccess: isConfirmed,
-    error: receiptError
-  } = useWaitForTransactionReceipt({ hash: writeContractData })
-
-  const isLoading = computed(() => isWritePending.value || isConfirming.value)
-
-  /**
-   * Invalidate specific Bank contract queries after successful transactions
-   */
-  const invalidateBankQueries = async (functionName: BankFunctionName) => {
-    if (!bankAddress.value) return
-
-    const bankQueryKey = {
-      address: bankAddress.value,
-      chainId: chainId.value
-    }
-
-    // Invalidate queries based on the function that was called
-    switch (functionName) {
-      case BANK_FUNCTION_NAMES.PAUSE:
-      case BANK_FUNCTION_NAMES.UNPAUSE:
-        // Invalidate paused status query
-        const key = {
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.PAUSED
-            }
-          ]
-        }
-        console.log("try to invalidate this key", key)
-        await queryClient.invalidateQueries(key)
-        break
-
-      case BANK_FUNCTION_NAMES.CHANGE_TIPS_ADDRESS:
-        // Invalidate tips address query
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.PAUSED
-            }
-          ]
-        })
-        break
-
-      case BANK_FUNCTION_NAMES.CHANGE_TIPS_ADDRESS:
-        // Invalidate tips address query
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.TIPS_ADDRESS
-            }
-          ]
-        })
-        break
-
-      case BANK_FUNCTION_NAMES.TRANSFER_OWNERSHIP:
-      case BANK_FUNCTION_NAMES.RENOUNCE_OWNERSHIP:
-        // Invalidate owner query
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.OWNER
-            }
-          ]
-        })
-        break
-
-      case BANK_FUNCTION_NAMES.CHANGE_TOKEN_ADDRESS:
-        // Invalidate all token-related queries
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.IS_TOKEN_SUPPORTED
-            }
-          ]
-        })
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            {
-              ...bankQueryKey,
-              functionName: BANK_FUNCTION_NAMES.SUPPORTED_TOKENS
-            }
-          ]
-        })
-        break
-
-      case BANK_FUNCTION_NAMES.TRANSFER:
-      case BANK_FUNCTION_NAMES.TRANSFER_TOKEN:
-      case BANK_FUNCTION_NAMES.SEND_TIP:
-      case BANK_FUNCTION_NAMES.SEND_TOKEN_TIP:
-      case BANK_FUNCTION_NAMES.PUSH_TIP:
-      case BANK_FUNCTION_NAMES.PUSH_TOKEN_TIP:
-      case BANK_FUNCTION_NAMES.DEPOSIT_TOKEN:
-        // Invalidate balance queries
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'balance',
-            bankQueryKey
-          ]
-        })
-        // Also invalidate all readContract queries for this bank address
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            bankQueryKey
-          ]
-        })
-        break
-
-      default:
-        // For any other function, invalidate all bank-related queries
-        await queryClient.invalidateQueries({
-          queryKey: [
-            'readContract',
-            bankQueryKey
-          ]
-        })
-        break
-    }
-  }
-
-  // Error handling
-  watch(writeError, (error) => {
-    if (error) {
-      console.error('Bank contract write error:', error)
-      addErrorToast(`Transaction failed: ${parseError(error)}`)
-    }
-  })
-
-  watch(receiptError, (error) => {
-    if (error) {
-      console.error('Bank transaction receipt error:', error)
-      addErrorToast(`Transaction confirmation failed: ${parseError(error)}`)
-    }
-  })
-
-  // Success handling with query invalidation
-  watch(isConfirmed, async (confirmed) => {
-    if (confirmed && receipt.value && currentFunctionName.value) {
-      addSuccessToast('Transaction confirmed successfully')
-      
-      // Invalidate queries based on the function that was executed
-      try {
-        await invalidateBankQueries(currentFunctionName.value)
-      } catch (error) {
-        console.warn('Could not invalidate queries for function:', currentFunctionName.value, error)
-        // Fallback: invalidate all bank queries
-        if (bankAddress.value) {
-          await queryClient.invalidateQueries({
-            queryKey: [
-              'readContract',
-              {
-                address: bankAddress.value,
-                chainId: chainId.value
-              }
-            ]
-          })
-        }
-      } finally {
-        // Clear the function name after processing
-        currentFunctionName.value = null
-      }
-    }
-  })
-
-  const executeWrite = async (
-    functionName: BankFunctionName, 
-    args: readonly unknown[] = [], 
-    value?: bigint
-  ) => {
-    if (!bankAddress.value) {
-      addErrorToast('Bank contract address not found')
-      return
-    }
-
-    if (!isValidBankFunction(functionName)) {
-      addErrorToast(`Invalid bank function: ${functionName}`)
-      return
-    }
-
-    // Store function name for query invalidation after transaction confirms
-    currentFunctionName.value = functionName
-
-    try {
-      await writeContract({
-        address: bankAddress.value,
-        abi: BankABI,
-        functionName,
-        args,
-        ...(value && { value })
-      })
-    } catch (error) {
-      console.error(`Failed to execute ${functionName}:`, error)
-      addErrorToast(`Failed to execute ${functionName}`)
-      // Clear function name on error
-      currentFunctionName.value = null
-    }
-  }
-
-  return {
-    isLoading, // Loading state of useWriteContract
-    isWritePending, // Loading state of useWriteContract & useWaitForTransactionReceipt
-    isConfirming, // Loading state of useWaitForTransactionReceipt
-    isConfirmed, // State of the transaction receipt 
-    writeContractData, // Write contract hash
-    receipt, // Receipt
-    executeWrite,
-    invalidateBankQueries // Expose for manual invalidation if needed
-  }
-}
-
-/**
- * Bank contract write functions - combines admin, transfers, and tipping
- */
-export function useBankWritesFunctions() {
-  const writes = useBankWrites()
-  const { addErrorToast } = useToastStore()
-
-  // Admin functions
-  const pauseContract = () => writes.executeWrite(BANK_FUNCTION_NAMES.PAUSE)
-  const unpauseContract = () => writes.executeWrite(BANK_FUNCTION_NAMES.UNPAUSE)
-
-  const changeTipsAddress = (newTipsAddress: Address) => {
-    if (!isAddress(newTipsAddress)) {
-      addErrorToast('Invalid tips address')
-      return
-    }
-    return writes.executeWrite(BANK_FUNCTION_NAMES.CHANGE_TIPS_ADDRESS, [newTipsAddress])
-  }
-
-  const changeTokenAddress = (symbol: string, newTokenAddress: Address) => {
-    if (!isAddress(newTokenAddress)) {
-      addErrorToast('Invalid token address')
-      return
-    }
-    if (!symbol.trim()) {
-      addErrorToast('Token symbol is required')
-      return
-    }
-    return writes.executeWrite(BANK_FUNCTION_NAMES.CHANGE_TOKEN_ADDRESS, [symbol, newTokenAddress])
-  }
-
-  const transferOwnership = (newOwner: Address) => {
-    if (!isAddress(newOwner)) {
-      addErrorToast('Invalid new owner address')
-      return
-    }
-    return writes.executeWrite(BANK_FUNCTION_NAMES.TRANSFER_OWNERSHIP, [newOwner])
-  }
-
-  const renounceOwnership = () => writes.executeWrite(BANK_FUNCTION_NAMES.RENOUNCE_OWNERSHIP)
-
-  // Transfer functions
-  const validateAmount = (amount: string) => {
-    if (!amount || parseFloat(amount) <= 0) {
-      addErrorToast('Invalid amount')
-      return false
-    }
-    return true
-  }
-
-  const validateAddress = (address: Address, label = 'address') => {
-    if (!isAddress(address)) {
-      addErrorToast(`Invalid ${label}`)
-      return false
-    }
-    return true
-  }
-
-  const depositToken = (tokenAddress: Address, amount: string) => {
-    if (!validateAddress(tokenAddress, 'token address') || !validateAmount(amount)) return
-    const amountInWei = parseEther(amount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.DEPOSIT_TOKEN, [tokenAddress, amountInWei])
-  }
-
-  const transferEth = (to: Address, amount: string) => {
-    if (!validateAddress(to, 'recipient address') || !validateAmount(amount)) return
-    const amountInWei = parseEther(amount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.TRANSFER, [to, amountInWei], amountInWei)
-  }
-
-  const transferToken = (tokenAddress: Address, to: Address, amount: string) => {
-    if (!validateAddress(tokenAddress, 'token address') ||
-      !validateAddress(to, 'recipient address') ||
-      !validateAmount(amount)) return
-    const amountInWei = parseEther(amount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.TRANSFER_TOKEN, [tokenAddress, to, amountInWei])
-  }
-
-  // Tipping functions
-  const validateTipParams = (addresses: Address[], amount: string, tokenAddress?: Address) => {
-    if (!addresses.length) {
-      addErrorToast('No recipients specified')
-      return false
-    }
-    if (addresses.some(addr => !isAddress(addr))) {
-      addErrorToast('One or more invalid addresses')
-      return false
-    }
-    if (!amount || parseFloat(amount) <= 0) {
-      addErrorToast('Invalid amount')
-      return false
-    }
-    if (tokenAddress && !isAddress(tokenAddress)) {
-      addErrorToast('Invalid token address')
-      return false
-    }
-    return true
-  }
-
-  const sendEthTip = (addresses: Address[], totalAmount: string) => {
-    if (!validateTipParams(addresses, totalAmount)) return
-    const amountInWei = parseEther(totalAmount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.SEND_TIP, [addresses, amountInWei], amountInWei)
-  }
-
-  const sendTokenTip = (addresses: Address[], tokenAddress: Address, totalAmount: string) => {
-    if (!validateTipParams(addresses, totalAmount, tokenAddress)) return
-    const amountInWei = parseEther(totalAmount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.SEND_TOKEN_TIP, [addresses, tokenAddress, amountInWei])
-  }
-
-  const pushEthTip = (addresses: Address[], totalAmount: string) => {
-    if (!validateTipParams(addresses, totalAmount)) return
-    const amountInWei = parseEther(totalAmount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.PUSH_TIP, [addresses, amountInWei], amountInWei)
-  }
-
-  const pushTokenTip = (addresses: Address[], tokenAddress: Address, totalAmount: string) => {
-    if (!validateTipParams(addresses, totalAmount, tokenAddress)) return
-    const amountInWei = parseEther(totalAmount)
-    return writes.executeWrite(BANK_FUNCTION_NAMES.PUSH_TOKEN_TIP, [addresses, tokenAddress, amountInWei])
-  }
-
-  return {
-    // Write state
-    ...writes,
-    // Admin functions
-    pauseContract,
-    unpauseContract,
-    changeTipsAddress,
-    changeTokenAddress,
-    transferOwnership,
-    renounceOwnership,
-    // Transfer functions
-    depositToken,
-    transferEth,
-    transferToken,
-    // Tipping functions
-    sendEthTip,
-    sendTokenTip,
-    pushEthTip,
-    pushTokenTip
-  }
-}
-
-/**
- * Main Bank contract composable - combines all functionality
- */
-export function useBankContract() {
-  const reads = useBankReads()
-  const writeFunctions = useBankWritesFunctions()
-
-  return {
-    ...reads,
-    ...writeFunctions
-  }
-}
+export * from './bank'
+export { default as useBankContract } from './bank'
